Way to much for a single Commit. So suit me.
Lost of Lion fixes, including the Find… fix
Fixing some NIBs
Some markdown autocomplete
and more
worst commit yet?

+++ EmptyView.h
@@ -17,11 +17,11 @@
 @interface EmptyView : NSView
 {
     IBOutlet NSTextField *labelText;
-	int lastNotesNumber;
+	NSInteger lastNotesNumber;
 	NSColor *bgCol;
 }
 
-- (void)setLabelStatus:(int)notesNumber;
+- (void)setLabelStatus:(NSInteger)notesNumber;
 - (void)setBackgroundColor:(NSColor *)inColor;
 
 @end
